feat: Switch to django-core
Make the wishlist a standalone django app which depends on django core

BREAKING CHANGE: The names of the database tables change. All existing
tables
with the new names will be overwriten. The new names are:
wishlist_wish, wishlist_group, wishlist_reservation and
wishlist_group_users. If the database is only used by the wishlist
everything is fine. Otherwise check if one of the tables exists already.